Remove some dead code
